dev-lang/php: Guard fix for #977105 behind 'fpm' use flag
Otherwise, removal of the directories will fail because they
are only pulled in by the FPM SAPI.

+++ dev-lang/php/php-8.4.23.ebuild
@@ -716,7 +716,9 @@ src_install() {
 
 	# Delete empty /var/run and /var/log directories.
 	# See bug #977105
-	rmdir "${D}/var/log" "${D}/var/run" || die
+	if use fpm ; then
+		rmdir "${D}/var/log" "${D}/var/run" || die
+	fi
 }
 
 src_test() {
